refactor(ComboBox): rename constants.ts to types.ts
